fix: Include final newline when writing to .gitignore
Leaving the file without a final newline in it makes it look ugly
for example when you run `cat .glab-cli/config/.env`.

Also drop the prepend of a newline is the ReadAndAppend method
to make it actually (only) append the data passed into it.
(This feels almost like a workaround for the previous missing trailing
newline, so should not be needed anymore now.)