fix: File content modification made a bit more safe and sane (#202)
* Use atomic replace in config.SetEnv

This commits makes the .env file be replaced atomically.

The details of this commit includes:
- bail out when reading existing file fails, instead of destroying
  existing config file.
- importing the github.com/google/renameio module
- dropping unused lines variable
- dropping fmt.Sprint of item
- avoid stripping final newlines in created file
- skip over empty lines (ie. when reading data from non-existant file,
  last item in range after splitting on final newline in file)

* fix: Include final newline when writing to .gitignore

Leaving the file without a final newline in it makes it look ugly
for example when you run `cat .glab-cli/config/.env`.

Also drop the prepend of a newline is the ReadAndAppend method
to make it actually (only) append the data passed into it.
(This feels almost like a workaround for the previous missing trailing
newline, so should not be needed anymore now.)