fix(config): use ioutil.WriteFile instead of renameio
As reported apparently renameio does not work on windows and there's
no obvious way to even make an atomic file overwrite on windows at all,
thus renameio is looking at dropping Windows support.

This change simply switches to ioutil.WriteFile which is not atomic
but the code should still be nicer than the previous one we replaced.

While at it also cut down on the readability part of the files
permissions as the config file could contain sensitive information
(like the token).

Fixes #210