feat: support "Run pipeline"
The API calls it "create pipeline", but the gitlab
web interface button to do this says "Run pipeline".

I named the command run as that's likely to be more familiar
to users.

With this you can now support starting a pipeline via
"glab pipeline run -b mybranch".

Note that setting Ref when calling the CreatePipeline API
seems required or we'll end up with an "Internal Server Error".
We thus do a best-effort attempt at figuring out the default
when the branch argument is not specified.