fix(mr): Readd ability to close multiple MRs
Ability to remove multiple MRs was there but it had gone missing when
implementing `MRFromArgs`.

Added a function for concurrently fetching multiple merge requests from
branches that, fallsback to the `MRFromArgs` if the argument count is
lower then 1. This is because the MRFromArgs, contains some specific
logic for handling cases where there are no arguments or user provided
0 in arguments.

Fixes #301