feat(commands/cmdutils): add initial value of NoAction
This fixes a nasty bug where our mr_create and issue_create code was
using this to skip confirmation.

```
// submit without prompting for non interactive mode
if !opts.IsInteractive {
    action = cmdutils.SubmitAction
}
```

This causes as bug as action was always started with the 0th value which
is cmdutils.SubmitAction, so not being interactive still caused a
confirmation prompt to appear.

with this change the 0th value is now cmdutils.NoAction and the code
above now works as intended, skipping the submission confirmation.