feat(commands/mr/mrutils): add OWNER:branch support
This adds support for querying Merge Requests via '[OWNER:]branch'.

Users can ask `mr view <branch>`, and it will look for a merge request
with the source branch of <branch>.

If it finds none then error out.
If it finds only one then it returns the merge request to the user.
If it finds more than one it then prompts the user to pick one of the
merge requests and returns the picked one.

If the user passes the optional OWNER: field then it also tries to it to
the username of the creator of the merge request.

Example: The user passes profclems:foo

It will search all merge requests for ones with the source branch `foo`

If it finds none then it errors out.
If it finds one it tries to match the Author.Username to `profclems`,
returning it if it matches or erroring out if it doesn't.
If it finds more than one, it checks if **any** of the merge requests
matches the Author.Username field and if matched it returns it otherwise
it errors out.