fix: remotes error on issue create
When an issue is created for a different repo which has no remote
in the current working git directory, it results in a `remote not found` error.
This happens because `cmdutils.LabelsPrompt` needs the remote name to cache the labels
in .git/config.

To fix that, we avoid caching the labels if there are no matching remotes.