refactor(internal/glrepo/resolver.go): remove code that is always false
This code is always false because of the following code that comes a bit
before.

```golang
if r.network == nil {
    resolveNetwork(r)
    if len(r.network) == 0 {
        return nil, errors.New("no GitLab Projects found from remotes")
    }
}
```

This code will always return err if there are no `gitlab.Project` in
r.network even after using `resolveNetwork()`, and the amount of values
in `repoNames` on how many values there are in r.network.

Thus the code we are removing is never actually reached, or rather, it
is in all situations, false.