chore(ci/test): switch test job from GitLab CI to GitHub action
Currently, GitLab CI does not run on forks per the discussion here https://gitlab.com/gitlab-org/gitlab/-/issues/5667.
However, `glab` needs the GitLab CI to run on forks in order to run tests.
Since this is not possible, we remove the test job from the GitLab CI to gitHub action.

GitHub action provides pull_request_target event which is not limited to secrets and runs on forks.