respect GITLAB_HOST when resolving remotes
When GITLAB_HOST or GITLAB_URI has been set, we resolve only local remotes that match the set hostname.
With this in place, a heterogeneous mix of hosts in a local remote file will not cause GITLAB_HOST to be ignored.

An error is returned if the hostname specified in GITLAB_HOST is not present in the local remotes.