fix(mr): Look for all MRs with `mr` commands
When looking for MR from arguments, we where looking only for open merge
requests. This doesn't allow commands like `glab mr reopen <branch>`.

In reality it didn't allow any commands to run on closed MRs using
branch names.

The only thing that is needed to consider that some commands can't run
on closed MRs like rebase or merge. There were checks for that in some
places, but not other.

Issue #620