fix(board): Don't display closed issues in columns
In web UI issues that are closed are moved into their own column. They
shouldn't belong in the main columns, if they are closed.

There is the thing that issues that don't belong in any of user
made boards, should be in their own boards.

Issue for adding the extra columns #665

Fixed #664