Fix method not allowed when merging MR
When merging MR, the --when-pipeline-succeeds flag is by default set to true.
Due to that, if a project does not have any pipeline set, results in error
`method not allowed`.
This commit fixes this issue by checking if the merge request has a pipeline set.
Also, on TTYs, users are prompted to confirm whether to merge when pipeline succeeds or not,
if the MR has a pipeline set.