Fix triggered jobs duration
This commit fixes a bug where a `glab ci status` would panic if there is
a job within that CI pipeline that is not triggered yet. In such a case,
the StartedAt field is not populated, which triggered a panic.

To fix this, we use `time.Now()` if `StartedAt` is nil, which means that
the job will be displayed with a duration of 0.

Fixes #727