add remoteName flag to 'create project' command
By default glab would add the created project's repository URL to the a
remote named 'origin', which fails when there already is such a remote.

This commit adds the option to specify a remote name for glab to add in
the local git repository, so that the user may specify a different one,
should 'origin' be already taken.

Closes #803.