mr: deprecate `mr for` command
The `mr for` command works just like create but accepts
issue ID as argument.

It creates a related merge request for the specified.
But for some reasons, it lacks some options available
when using `mr create`.

This change deprecates the `mr for` command and
adds additional options to `mr create` to perform
be able to create related merge request for issues