fix(api): Change from using retry request to option variable
The remove variable API didn't have option to set the enviorment scope,
so there was an workaround by hooking into the request retry option, and
setting the scope option by hand there.

This was fixed in the latest release of go-gitlab and now this can be
done simply by passing in a option object.